[implemented] creating new comment and getting all comments

+++ src/main/java/com/aspd/collegeCommunityPortal/beans/response/PostResponseViewList.java
@@ -10,6 +10,6 @@ import java.util.List;
 public class PostResponseViewList {
     private List<PostResponseView> postResponseViews;
     private int pageNo;
-    private int totalNoOfPost;
+    private long totalNoOfPost;
     private int totalPages;
 }
